requirements.txt
streamlit
langchain==0.1.16
langchain-core==0.1.52
langchainhub==0.1.21
langchain-community==0.0.36
langchain-groq==0.0.1
groq
faiss-cpu
PyPDF2
python-dotenv
